Добавил закладки через React Router

+++ src/App.js
@@ -1,27 +1,40 @@
 import React from "react";
+import { Routes, Route } from "react-router-dom";
 import axios from "axios";
-import Card from "./components/Card";
+import Home from "./pages/Home";
+import Favorites from "./pages/Favorites";
 import Header from "./components/Header";
 import Cart from "./components/Cart";
 
 function App() {
     const [items, setItems] = React.useState([]);
     const [cartItems, setCartItems] = React.useState([]);
+    const [favoriteItems, setFavoriteItems] = React.useState([]);
     const [searchItem, setSearchItems] = React.useState();
     const [toggleCart, setToggleCart] = React.useState(false);
 
     const handlerAddItemCart = (card) => {
-        axios.post(
-            "https://62efc45857311485d127eb48.mockapi.io/cartItems",
-            card
-        );
         setCartItems([...cartItems, card]);
     };
 
-    const removeItemCart = (id, title) => {
-        axios.delete(
-            `https://62efc45857311485d127eb48.mockapi.io/cartItems/${id}`
-        );
+    const handlerAddToFavorite = (card) => {
+        if (favoriteItems.find((obj) => obj.id === card.id)) {
+            axios.delete(
+                `https://62efc45857311485d127eb48.mockapi.io/favorite/${card.id}`
+            );
+            setFavoriteItems((prev) =>
+                prev.filter((item) => item.id !== card.id)
+            );
+        } else {
+            axios.post(
+                "https://62efc45857311485d127eb48.mockapi.io/favorite",
+                card
+            );
+            setFavoriteItems((prev) => [...prev, card]);
+        }
+    };
+
+    const removeItemCart = (title) => {
         setCartItems((prev) => prev.filter((item) => item.title !== title));
     };
 
@@ -37,9 +50,10 @@ function App() {
         axios
             .get("https://62efc45857311485d127eb48.mockapi.io/items")
             .then((res) => setItems(res.data));
+
         axios
-            .get("https://62efc45857311485d127eb48.mockapi.io/cartItems")
-            .then((res) => setCartItems(res.data));
+            .get("https://62efc45857311485d127eb48.mockapi.io/favorite")
+            .then((res) => setFavoriteItems(res.data));
     }, []);
 
     return (
@@ -47,50 +61,39 @@ function App() {
             {toggleCart && (
                 <Cart
                     items={cartItems}
-                    deleteCard={(id, title) => removeItemCart(id, title)}
+                    removeItem={removeItemCart}
                     closeCart={() => setToggleCart(false)}
                 />
             )}
+
             <Header showCart={() => setToggleCart(true)} />
-            <div className="content p-40">
-                <div className="d-flex mb-40 align-center justify-between">
-                    <h1>Все кроссовки</h1>
-                    <div className="search-block d-flex">
-                        <img src="/img/search.svg" alt="search" />
-                        <input
-                            onChange={changeSearchInput}
-                            value={searchItem}
-                            placeholder="Поиск..."
+
+            <Routes>
+                <Route
+                    path="/"
+                    exact
+                    element={
+                        <Home
+                            items={items}
+                            searchItem={searchItem}
+                            changeSearchInput={changeSearchInput}
+                            clearSeacrhInput={clearSeacrhInput}
+                            handlerAddItemCart={handlerAddItemCart}
+                            handlerAddToFavorite={handlerAddToFavorite}
                         />
-                        {searchItem && (
-                            <img
-                                src="/img/btn_remove.svg"
-                                alt="clear"
-                                onClick={clearSeacrhInput}
-                            />
-                        )}
-                    </div>
-                </div>
-                <div className="cards_container d-flex flex-wrap">
-                    {items
-                        /* .filter((item) =>
-                            item.title
-                                .toLowerCase()
-                                .includes(
-                                    searchItem && searchItem.toLowerCase()
-                                )
-                        ) */
-                        .map((card, index) => (
-                            <Card
-                                key={index}
-                                title={card.title}
-                                price={card.price}
-                                imageUrl={card.imageUrl}
-                                addItemCart={() => handlerAddItemCart(card)}
-                            />
-                        ))}
-                </div>
-            </div>
+                    }
+                ></Route>
+                <Route
+                    path="/favorites"
+                    exact
+                    element={
+                        <Favorites
+                            items={favoriteItems}
+                            handlerAddToFavorite={handlerAddToFavorite}
+                        />
+                    }
+                ></Route>
+            </Routes>
         </div>
     );
 }
